Feat/#3.1 add distance for calculation

+++ Route256/ApiModels/V2/GetHistory/GetHistoryResponse.cs
@@ -3,4 +3,4 @@ namespace Route256.ApiModels.V2.GetHistory;
 public record GetHistoryResponse(CargoHistoryResponse[] CargoHistory);
 
 
-public record CargoHistoryResponse(decimal Volume, decimal Weight, decimal Price);
+public record CargoHistoryResponse(double Volume, double Weight, decimal Price);
